new basket design that is more mobile friendly
