+++ public/components/BasketSummary.styl
@@ -3,7 +3,7 @@
     display flex
     flex-direction column
     justify-content center
-    margin 0 10px
+    margin 15px 10px
     min-height 75%
 
 .container
@@ -20,9 +20,6 @@
 
 .actions
     text-align center
-
-    *:not(:first-child)
-        margin-left 5px
 
 @media (max-width 736px)
     .root
