+++ public/components/BasketSummaryListItem.styl
@@ -1,0 +1,27 @@
+$avatar-width = 60px
+$top-bottom-margin = 12px
+
+.root
+    align-items center
+    display flex
+    padding 0 16px
+
+.avatar
+    border-radius 50%
+    height $avatar-width
+    margin $top-bottom-margin 16px $top-bottom-margin 0
+    width $avatar-width
+
+.textContainer
+    flex 1
+    margin-right 16px
+
+.name
+    font-size 1.2em
+    margin $top-bottom-margin 0
+
+.total
+    font-size 1em
+
+.quantity
+    margin 0 10px
