--- public/components/BasketSummaryRow.styl
@@ -1,2 +1,0 @@
-.quantity
-    margin 0 10px
